# Minor performance improvement for UriBuildingContext.ToPathString() w…