Minor performance improvement for `UriBuildingContext.ToPathString()` with zero-length path.
Added tests for UriBuildingContext.ToPathString().

**Before:**

|                                                               Method |      Mean |     Error |    StdDev |          Op/s |  Gen 0 | Allocated |
|--------------------------------------------------------------------- |----------:|----------:|----------:|--------------:|-------:|----------:|
|                      ToPathString_PathEmpty_AppendTrailingSlashFalse |  5.828 ns | 0.0271 ns | 0.0253 ns | 171,581,171.0 |      - |       0 B |
|                       ToPathString_PathEmpty_AppendTrailingSlashTrue |  5.792 ns | 0.0141 ns | 0.0132 ns | 172,653,158.9 |      - |       0 B |
**After:**

|                                                               Method |      Mean |     Error |    StdDev |          Op/s |  Gen 0 | Allocated |
|--------------------------------------------------------------------- |----------:|----------:|----------:|--------------:|-------:|----------:|
|                      ToPathString_PathEmpty_AppendTrailingSlashFalse |  1.765 ns | 0.0031 ns | 0.0028 ns | 566,483,501.4 |      - |       0 B |
|                       ToPathString_PathEmpty_AppendTrailingSlashTrue |  1.760 ns | 0.0098 ns | 0.0092 ns | 568,144,661.1 |      - |       0 B |

Benchmark is available [here](https://gist.github.com/drieseng/1958a4755d08c77b231055172a568324).